+++ example/lib/core/common/words/words.dart
@@ -1,17 +1,12 @@
 import 'package:easy_localization/easy_localization.dart';
-import 'package:res_generator_annotation/res_generator_annotation.dart';
 
 part 'words.g.dart';
 
-@ResWordsAnnotation(
-  className: 'Words',
-  assetsDirectory: 'assets/tr/',
-  supportedLocales: ['uz', 'en'],
-  targetLocale: 'uz',
-)
+/// You can change code and method names
+/// use any package for example easy_localization
+/// use Words.hello.str() or str(Words.hello)
 extension MyString on String {
-  String genTr() {
-    //use any package for example easy_localization
-    return tr(this);
-  }
+  String str() => this.tr();
 }
+
+String str(String key) => key.tr();