+++ generators/CHANGELOG.md
@@ -1,3 +1,9 @@
-## 0.0.4
+## 0.1.0
+* remove dependencies of res_generator_annotations and build_runner packages
+* add custom generator `dart run res_generator:generate`
 
-* TODO: Describe initial release.
+## 0.0.7
+* Generate with res_generator_annotations and build_runner packages
+
+## 0.0.1
+* Initial release.